+++ src/Narvalo.Brouillons/Internal/Check.cs
@@ -16,7 +16,6 @@ namespace Narvalo.Internal
         /// Asserts that the specified type parameter is an enumeration.
         /// </summary>
         /// <typeparam name="T">The type to be checked.</typeparam>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]
@@ -31,7 +30,6 @@ namespace Narvalo.Internal
         /// Asserts that the specified type parameter is a value type.
         /// </summary>
         /// <typeparam name="T">The type to be checked.</typeparam>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]