+++ src/Narvalo.Common/Globalization/DefaultCurrencyProvider.g.cs
@@ -26,7 +26,7 @@ namespace Narvalo.Globalization
         /// </summary>
         /// <remarks>
         /// The list is built out of data found on the "Association Suisse de Normalisation"
-        /// <seealso href="http://www.currency-iso.org/en/home.html">website</seealso>.
+        /// <seealso cref="!:http://www.currency-iso.org/en/home.html">website</seealso>.
         /// </remarks>
         /// <value>The list of currently active currencies.</value>
         internal static IEnumerable<CurrencyInfo> CurrentCurrencies
@@ -1411,7 +1411,7 @@ namespace Narvalo.Globalization
         /// </summary>
         /// <remarks>
         /// The list is built out of data found on the "Association Suisse de Normalisation"
-        /// <seealso href="http://www.currency-iso.org/en/home.html">website</seealso>.
+        /// <seealso cref="!:http://www.currency-iso.org/en/home.html">website</seealso>.
         /// </remarks>
         /// <value>The list of discontinued currencies.</value>
         internal static IEnumerable<CurrencyInfo> LegacyCurrencies