+++ src/Narvalo.Core/Internal/FatalPreconditionException.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
+
+namespace Narvalo.Internal
+{
+    using System;
+    using System.Diagnostics.CodeAnalysis;
+
+    [SuppressMessage("Microsoft.Design", "CA1064:ExceptionsShouldBePublic",
+    Justification = "[Intentionally] This is an unrecoverable exception, thrown when a supposedly impossible situation happened.")]
+    [SuppressMessage("Gendarme.Rules.Exceptions", "MissingExceptionConstructorsRule",
+        Justification = "[Intentionally] This exception can not be initialized outside this assembly.")]
+    internal sealed class FatalPreconditionException : Exception
+    {
+        public FatalPreconditionException(string message) : base(message) { }
+    }
+}