+++ src/Narvalo.Core/Promise.cs
@@ -4,29 +4,32 @@ namespace Narvalo
 {
     using System;
     using System.Diagnostics;
-    using System.Diagnostics.CodeAnalysis;
     using System.Diagnostics.Contracts;
+
+    using Narvalo.Internal;
 
     /// <summary>
     /// Provides helper methods to state promises and to validate them right away.
     /// </summary>
     /// <remarks>
-    /// <para>The methods WON'T be recognized as parameter validators by FxCop.</para>
+    /// <para>The methods WON'T be recognized by FxCop as parameter validators
+    /// against <see langword="null"/> value.</para>
     /// <para>The methods will be recognized as Contract Abbreviator methods.</para>
     /// <para>If a promise does not hold, an unrecoverable exception is thrown.</para>
-    /// <para>This class MUST NOT be used in place of proper validation routines of public 
+    /// <para>This class MUST NOT be used in place of proper validation routines of public
     /// arguments but rather be reserved for internal sanity checking. Be wise.
     /// Personally, I can only see three situations where these helpers make sense:
     /// <list type="bullet">
     /// <item>for private methods where you have full control of all possible callers.</item>
     /// <item>for protected overriden methods in a sealed class when the base method does
-    /// not declare any contract AND when you know for certain that all callers will 
+    /// not declare any contract AND when you know for certain that all callers will
     /// satisfy the condition.</item>
     /// <item>and exceptionally for a few internal methods when you have achieved complete
     /// Code Contracts coverage.</item>
     /// </list>
     /// </para>
     /// </remarks>
+    [DebuggerStepThrough]
     public static class Promise
     {
         /// <summary>
@@ -34,12 +37,11 @@ namespace Narvalo
         /// </summary>
         /// <remarks>
         /// <para>All methods called within the condition must be pure.</para>
-        /// <para>All members mentioned in the condition must be at least as visible as the method 
+        /// <para>All members mentioned in the condition must be at least as visible as the method
         /// in which they appear.</para>
         /// </remarks>
         /// <param name="testCondition">The conditional expression to evaluate.</param>
         /// <param name="rationale">The rationale for the promise.</param>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]
@@ -49,7 +51,7 @@ namespace Narvalo
 
             if (!testCondition)
             {
-                throw new FailedPromiseException("A promise did not hold: " + rationale);
+                throw new FatalPreconditionException("A promise did not hold: " + rationale);
             }
         }
 
@@ -59,7 +61,6 @@ namespace Narvalo
         /// <typeparam name="T">The type of <paramref name="value"/>.</typeparam>
         /// <param name="value">The argument to check.</param>
         /// <param name="rationale">The rationale for the promise.</param>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]
@@ -69,7 +70,7 @@ namespace Narvalo
 
             if (value == null)
             {
-                throw new FailedPromiseException("The parameter value is null: " + rationale);
+                throw new FatalPreconditionException("The parameter value is null: " + rationale);
             }
         }
 
@@ -78,7 +79,6 @@ namespace Narvalo
         /// </summary>
         /// <param name="value">The argument to check.</param>
         /// <param name="rationale">The rationale for the promise.</param>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]
@@ -88,7 +88,7 @@ namespace Narvalo
 
             if (String.IsNullOrEmpty(value))
             {
-                throw new FailedPromiseException("The parameter value is null or empty: " + rationale);
+                throw new FatalPreconditionException("The parameter value is null or empty: " + rationale);
             }
         }
 
@@ -98,7 +98,6 @@ namespace Narvalo
         /// </summary>
         /// <param name="value">The argument to check.</param>
         /// <param name="rationale">The rationale for the promise.</param>
-        [DebuggerHidden]
         [ContractAbbreviator]
         [Conditional("DEBUG")]
         [Conditional("CONTRACTS_FULL")]
@@ -108,18 +107,9 @@ namespace Narvalo
 
             if (String.IsNullOrWhiteSpace(value))
             {
-                throw new FailedPromiseException(
+                throw new FatalPreconditionException(
                     "The parameter value is null or empty, or consists only of white-space characters: " + rationale);
             }
-        }
-
-        [SuppressMessage("Microsoft.Design", "CA1064:ExceptionsShouldBePublic",
-            Justification = "[Intentionally] This is an unrecoverable exception, thrown when a supposedly impossible situation happened.")]
-        [SuppressMessage("Gendarme.Rules.Exceptions", "MissingExceptionConstructorsRule",
-            Justification = "[Intentionally] This exception can not be initialized outside this assembly.")]
-        private sealed class FailedPromiseException : Exception
-        {
-            public FailedPromiseException(string message) : base(message) { }
         }
     }
 }